+++ App/xcbuddykit/Sources/Generator/ProjectFileElements.swift
@@ -58,14 +58,28 @@ class ProjectFileElements {
     func targetFiles(target: Target) -> Set<AbsolutePath> {
         var files = Set<AbsolutePath>()
         target.buildPhases.forEach { buildPhase in
+            // Sources
             if let sourcesBuildPhase = buildPhase as? SourcesBuildPhase {
-                files.formUnion(sourcesBuildPhase.buildFiles.files)
+                files.formUnion(sourcesBuildPhase.buildFiles.flatMap({ $0.paths }))
+
+                // Resources
             } else if let resourcesBuildPhase = buildPhase as? ResourcesBuildPhase {
-                files.formUnion(resourcesBuildPhase.buildFiles.files)
+                resourcesBuildPhase.buildFiles.forEach { buildFile in
+
+                    // Normal resources
+                    if let resourceBuildFile = buildFile as? ResourcesBuildFile {
+                        files.formUnion(resourceBuildFile.paths)
+
+                        // Core Data model resoureces
+                    } else if let coreDataModelBuildFile = buildFile as? CoreDataModelBuildFile {
+                        files.insert(coreDataModelBuildFile.path)
+                        files.formUnion(coreDataModelBuildFile.versions)
+                    }
+                }
+
+                // Headers
             } else if let headersBuildPhase = buildPhase as? HeadersBuildPhase {
-                files.formUnion(headersBuildPhase.public.files)
-                files.formUnion(headersBuildPhase.project.files)
-                files.formUnion(headersBuildPhase.private.files)
+                files.formUnion(headersBuildPhase.buildFiles.flatMap({ $0.paths }))
             }
         }
         // Support files
@@ -116,7 +130,7 @@ class ProjectFileElements {
         let closestRelativeAbsolutePath = sourceRootPath.appending(closestRelativeRelativePath)
 
         // Add the first relative element.
-        let firstElement = addComponent(relative: closestRelativeRelativePath, from: sourceRootPath, toGroup: groups.project, objects: objects)
+        let firstElement = addElement(relativePath: closestRelativeRelativePath, from: sourceRootPath, toGroup: groups.project, objects: objects)
 
         // If it matches the file that we are adding or it's not a group we can exit.
         if (closestRelativeAbsolutePath == path) || !(firstElement.element is PBXGroup) {
@@ -129,13 +143,169 @@ class ProjectFileElements {
 
         path.relative(to: lastPath).components.forEach { component in
             if lastGroup == nil { return }
-            let element = addComponent(relative: RelativePath(component),
-                                       from: lastPath,
-                                       toGroup: lastGroup!,
-                                       objects: objects)
+            let element = addElement(relativePath: RelativePath(component),
+                                     from: lastPath,
+                                     toGroup: lastGroup!,
+                                     objects: objects)
             lastGroup = element.element as? PBXGroup
             lastPath = element.path
         }
+    }
+
+    // MARK: - Fileprivate
+
+    /// Adds a new file or group to an existing group.
+    ///
+    /// - Parameters:
+    ///   - relative: relative path from the group that is passed.
+    ///   - from: absolute path of the group that is passed.
+    ///   - toGroup: group where the file/group should be added.
+    ///   - objects: project objects.
+    @discardableResult func addElement(relativePath: RelativePath,
+                                       from: AbsolutePath,
+                                       toGroup: PBXGroup,
+                                       objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
+        let absolutePath = from.appending(relativePath)
+        if elements[absolutePath] != nil {
+            return (element: elements[absolutePath]!, path: from.appending(relativePath))
+        }
+
+        // If the path is ../../xx we specify the name
+        // to prevent Xcode from using that as a name.
+        var name: String?
+        let components = relativePath.components
+        if components.count != 1 {
+            name = components.last!
+        }
+
+        // Add the file element
+        if isVariantGroup(path: absolutePath) {
+            return addVariantGroupElement(from: from,
+                                          folderAbsolutePath: absolutePath,
+                                          folderRelativePath: relativePath,
+                                          name: name,
+                                          toGroup: toGroup,
+                                          objects: objects)
+        } else if isVersionGroup(path: absolutePath) {
+            return addVersionGroupElement(from: from,
+                                          folderAbsolutePath: absolutePath,
+                                          folderRelativePath: relativePath,
+                                          name: name,
+                                          toGroup: toGroup,
+                                          objects: objects)
+        } else if isGroup(path: absolutePath) {
+            return addGroupElement(from: from,
+                                   folderAbsolutePath: absolutePath,
+                                   folderRelativePath: relativePath,
+                                   name: name,
+                                   toGroup: toGroup,
+                                   objects: objects)
+        } else {
+            return addFileElement(from: from,
+                                  fileAbsolutePath: absolutePath,
+                                  fileRelativePath: relativePath,
+                                  name: name,
+                                  toGroup: toGroup,
+                                  objects: objects)
+        }
+    }
+
+    /// Adds a variant group element.
+    ///
+    /// - Parameters:
+    ///   - from: absolute path of the group the group is being added to.
+    ///   - folderAbsolutePath: folder absolute path.
+    ///   - folderRelativePath: folder path relative to the group absolute path.
+    ///   - name: element name.
+    ///   - toGroup: group where the new group will be added.
+    ///   - objects: Xcode project objects.
+    /// - Returns: added group.
+    func addVariantGroupElement(from: AbsolutePath,
+                                folderAbsolutePath: AbsolutePath,
+                                folderRelativePath: RelativePath,
+                                name: String?,
+                                toGroup: PBXGroup,
+                                objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
+        let group = PBXVariantGroup(children: [], sourceTree: .group, name: name, path: folderRelativePath.asString)
+        let reference = objects.addObject(group)
+        toGroup.children.append(reference)
+        elements[folderAbsolutePath] = group
+        return (element: group, path: from.appending(folderRelativePath))
+    }
+
+    /// Adds a version group element.
+    ///
+    /// - Parameters:
+    ///   - from: absolute path of the group the group is being added to.
+    ///   - folderAbsolutePath: folder absolute path.
+    ///   - folderRelativePath: folder path relative to the group absolute path.
+    ///   - name: element name.
+    ///   - toGroup: group where the new group will be added.
+    ///   - objects: Xcode project objects.
+    /// - Returns: added group.
+    func addVersionGroupElement(from: AbsolutePath,
+                                folderAbsolutePath: AbsolutePath,
+                                folderRelativePath: RelativePath,
+                                name: String?,
+                                toGroup: PBXGroup,
+                                objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
+        let versionGroupType = Xcode.filetype(extension: folderRelativePath.extension!)
+        let group = XCVersionGroup(currentVersion: nil,
+                                   path: folderRelativePath.asString,
+                                   name: name,
+                                   sourceTree: .group,
+                                   versionGroupType: versionGroupType)
+        let reference = objects.addObject(group)
+        toGroup.children.append(reference)
+        elements[folderAbsolutePath] = group
+        return (element: group, path: from.appending(folderRelativePath))
+    }
+
+    /// Adds a normal group element.
+    ///
+    /// - Parameters:
+    ///   - from: absolute path of the group the group is being added to.
+    ///   - folderAbsolutePath: folder absolute path.
+    ///   - folderRelativePath: folder path relative to the group absolute path.
+    ///   - name: element name.
+    ///   - toGroup: group where the new group will be added.
+    ///   - objects: Xcode project objects.
+    /// - Returns: added group.
+    func addGroupElement(from: AbsolutePath,
+                         folderAbsolutePath: AbsolutePath,
+                         folderRelativePath: RelativePath,
+                         name: String?,
+                         toGroup: PBXGroup,
+                         objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
+        let group = PBXGroup(children: [], sourceTree: .group, name: name, path: folderRelativePath.asString)
+        let reference = objects.addObject(group)
+        toGroup.children.append(reference)
+        elements[folderAbsolutePath] = group
+        return (element: group, path: from.appending(folderRelativePath))
+    }
+
+    /// Adds a file element.
+    ///
+    /// - Parameters:
+    ///   - from: absolute path of the group the file is being added to.
+    ///   - fileAbsolutePath: file absolute path.
+    ///   - fileRelativePath: file path relative to the group absolute path.
+    ///   - name: element name.
+    ///   - toGroup: group where the file will be added.
+    ///   - objects: Xcode project objects.
+    /// - Returns: added file.
+    func addFileElement(from: AbsolutePath,
+                        fileAbsolutePath: AbsolutePath,
+                        fileRelativePath: RelativePath,
+                        name: String?,
+                        toGroup: PBXGroup,
+                        objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
+        let lastKnownFileType = Xcode.filetype(extension: fileAbsolutePath.extension!)
+        let file = PBXFileReference(sourceTree: .group, name: name, lastKnownFileType: lastKnownFileType, path: fileRelativePath.asString)
+        let reference = objects.addObject(file)
+        toGroup.children.append(reference)
+        elements[fileAbsolutePath] = file
+        return (element: file, path: from.appending(fileRelativePath))
     }
 
     /// Returns the group at the given path if it's been added to the file elements object.
@@ -154,49 +324,28 @@ class ProjectFileElements {
         return elements[path] as? PBXFileReference
     }
 
-    // MARK: - Fileprivate
-
-    /// Adds a new file or group to an existing group.
+    /// Returns true if the path is a version group.
     ///
-    /// - Parameters:
-    ///   - relative: relative path from the group that is passed.
-    ///   - from: absolute path of the group that is passed.
-    ///   - toGroup: group where the file/group should be added.
-    ///   - objects: project objects.
-    @discardableResult func addComponent(relative: RelativePath,
-                                         from: AbsolutePath,
-                                         toGroup: PBXGroup,
-                                         objects: PBXObjects) -> (element: PBXFileElement, path: AbsolutePath) {
-        let absolutePath = from.appending(relative)
-        if elements[absolutePath] != nil {
-            return (element: elements[absolutePath]!, path: from.appending(relative))
-        }
+    /// - Parameter path: path.
+    /// - Returns: true if the path is a version group.
+    func isVersionGroup(path: AbsolutePath) -> Bool {
+        return path.extension == "xcdatamodeld"
+    }
 
-        // If the path is ../../xx we specify the name
-        // to prevent Xcode from using that as a name.
-        var name: String?
-        let components = relative.components
-        if components.count != 1 {
-            name = components.last!
-        }
+    /// Returns true if the group is a variant group.
+    ///
+    /// - Parameter path: path.
+    /// - Returns: true if the group is a variant group.
+    func isVariantGroup(path: AbsolutePath) -> Bool {
+        return path.extension == "lproj"
+    }
 
-        // Folder
-        if relative.extension == nil {
-            let group = PBXGroup(children: [], sourceTree: .group, name: name, path: relative.asString)
-            let reference = objects.addObject(group)
-            toGroup.children.append(reference)
-            elements[absolutePath] = group
-            return (element: group, path: from.appending(relative))
-
-            // File
-        } else {
-            let lastKnownFileType = Xcode.filetype(extension: absolutePath.extension!)
-            let file = PBXFileReference(sourceTree: .group, name: name, lastKnownFileType: lastKnownFileType, path: relative.asString)
-            let reference = objects.addObject(file)
-            toGroup.children.append(reference)
-            elements[absolutePath] = file
-            return (element: file, path: from.appending(relative))
-        }
+    /// Returns true if the path should be a group.
+    ///
+    /// - Parameter path: path.
+    /// - Returns: true if the path should be represented as a group.
+    func isGroup(path: AbsolutePath) -> Bool {
+        return !isVariantGroup(path: path) && !isVersionGroup(path: path) && path.extension == nil
     }
 
     /// Returns the relative path of the closest relative element to the source root path.