+++ Tests/ProjectDescriptionTests/TargetTests.swift
@@ -13,9 +13,9 @@ final class TargetTests: XCTestCase {
                              dependencies: [.framework(path: "path")],
                              settings: Settings(debug: .settings([:], xcconfig: "debug.xcconfig"),
                                                 release: .settings([:], xcconfig: "release.xcconfig")),
-                             buildPhases: [.headers()])
+                             buildPhases: [.headers([])])
         let json = subject.toJSON()
-        let expected = "{\"build_phases\": [{\"private\": [], \"project\": [], \"public\": [], \"type\": \"headers\"}], \"bundle_id\": \"bundle_id\", \"dependencies\": [{\"path\": \"path\", \"type\": \"framework\"}], \"entitlements\": \"entitlements\", \"info_plist\": \"info.plist\", \"name\": \"name\", \"platform\": \"ios\", \"product\": \"app\", \"settings\": {\"base\": {}, \"debug\": {\"settings\": {}, \"xcconfig\": \"debug.xcconfig\"}, \"release\": {\"settings\": {}, \"xcconfig\": \"release.xcconfig\"}}}"
+        let expected = "{\"build_phases\": [{\"files\": [], \"type\": \"headers\"}], \"bundle_id\": \"bundle_id\", \"dependencies\": [{\"path\": \"path\", \"type\": \"framework\"}], \"entitlements\": \"entitlements\", \"info_plist\": \"info.plist\", \"name\": \"name\", \"platform\": \"ios\", \"product\": \"app\", \"settings\": {\"base\": {}, \"debug\": {\"settings\": {}, \"xcconfig\": \"debug.xcconfig\"}, \"release\": {\"settings\": {}, \"xcconfig\": \"release.xcconfig\"}}}"
         XCTAssertEqual(json.toString(), expected)
     }
 }